Add initializeCharacterPosition to MapData type
Now, our intention is to add new maps in order to implement the teleport
action. So we include all the logic to initialize the app in a function
called initializeMap. Also, we need a position to start rendering the
character (before character position was hardcoded), so we add the
initializeCharacterPosition.